refactor(forge): split GitLab pipeline methods into gitlab_pipelines.go (#1025)
Pure same-package code motion: move LatestPipeline, LatestMRPipeline, ListPipelineJobs, JobLogTail, ProjectCIConfigPath, and the toPipeline helper out of gitlab.go into a new gitlab_pipelines.go, for parity with github_pipelines.go and forgejo_pipelines.go. No behaviour, signature, or doc-comment change. goimports dropped now-unused io and strconv from gitlab.go. No positional-comment fixes were needed.
